Rename topics middleware for consistency
