+++ lib/middleware/messages/member/topics/posts/photo/post-create.js
@@ -34,6 +34,7 @@ module.exports = function createPhotoPost() {
 
       return await helpers.replies.completedPhotoPost(req, res);
     } catch (err) {
+      // TODO: Prompt user to send a diffferent photo if Rogue returns error for image file size.
       return helpers.sendErrorResponse(res, err);
     }
   };
